refactor(preference): rename appearance store and files to preference for clarity
