+++ src/entrypoints/options/views/PreferenceView.vue
